Replace float with double. This is needed because voxels should technically be infinite, which can simply break the "float" limit, so "double" makes this limit much higher and we can safely generate literally infinite worlds!

+++ include/voxelius/globals.hh
@@ -9,9 +9,9 @@
 
 namespace globals
 {
-    extern float frame_time;
-    extern float current_time;
-    extern float last_time;
+    extern double frame_time;
+    extern double current_time;
+    extern double last_time;
 }
 
 #endif
