+++ src/globals.cc
@@ -7,7 +7,7 @@
 
 namespace globals
 {
-    float frame_time = 0.0f;
-    float current_time = 0.0f;
-    float last_time = 0.0f;
+    double frame_time = 0.0;
+    double current_time = 0.0;
+    double last_time = 0.0;
 }
